+++ FirewallProfile.ps1
@@ -47,7 +47,7 @@ Set-NetFirewallProfile -Profile Public -PolicyStore $PolicyStore `
 	-NotifyOnListen True -EnableStealthModeForIPsec True -AllowUnicastResponseToMulticast False `
 	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes 1024 `
 	-AllowUserApps NotConfigured -AllowUserPorts NotConfigured `
-	-LogFileName "%SystemRoot%\System32\LogFiles\Firewall\pfirewall.log" @Logs
+	-LogFileName "%SystemRoot%\System32\LogFiles\Firewall\PublicFirewall.log" @Logs
 
 # Setting up profile seem to be slow, tell user what is going on
 Write-Information -Tags "User" -MessageData "INFO: Setting up private firewall profile..." @Logs
@@ -58,7 +58,7 @@ Set-NetFirewallProfile -Profile Private -PolicyStore $PolicyStore `
 	-NotifyOnListen True -EnableStealthModeForIPsec True -AllowUnicastResponseToMulticast True `
 	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes 1024 `
 	-AllowUserApps NotConfigured -AllowUserPorts NotConfigured `
-	-LogFileName "%SystemRoot%\System32\LogFiles\Firewall\pfirewall.log" @Logs
+	-LogFileName "%SystemRoot%\System32\LogFiles\Firewall\PrivateFirewall.log" @Logs
 
 # Setting up profile seem to be slow, tell user what is going on
 Write-Information -Tags "User" -MessageData "INFO: Setting up domain firewall profile..." @Logs
@@ -69,6 +69,6 @@ Set-NetFirewallProfile -Profile Domain -PolicyStore $PolicyStore `
 	-NotifyOnListen True -EnableStealthModeForIPsec True -AllowUnicastResponseToMulticast True `
 	-LogAllowed False -LogBlocked True -LogIgnored True -LogMaxSizeKilobytes 1024 `
 	-AllowUserApps NotConfigured -AllowUserPorts NotConfigured `
-	-LogFileName "%SystemRoot%\System32\LogFiles\Firewall\pfirewall.log" @Logs
+	-LogFileName "%SystemRoot%\System32\LogFiles\Firewall\DomainFirewall.log" @Logs
 
 Update-Logs